RDVAULT-11: Set default ArchiveStores via configuration
Added support for configuring the default ArchiveStores set up in
`InitialiseDatabase` via the `default.properties` file.

Currently supports:
- local file
- S3, and
- Tivoli

Beans representing each of the above ArchiveStores are defined in 
`datavault-broker-servlet.xml` and then a `List` of these is injected 
into the `InitialiseDatabase` class.

Each ArchiveStore can be configured with an `enabled` property and it is
This that determines whether the `ArchiveStore` is persisted to the 
database. This new property has been added to the `ArchiveStore` class
and an additional class constructor signature including this property
has been defined.

The `archiveDir` property previously injected into `InitialiseDatabase`
has been removed as this property was relevant only to the local file 
ArchiveStore implementation and is now defined in the 
`localFileArchiveStore` bean definition.