RDVAULT-11: Add tests for `InitialiseDatabase` class
Added a few tests to check the functionality of the 
`initialiseArchiveStores` method in the `InitialiseDatabase` class.

Altered the `pom.xml` file for `datavault-broker` to look for test with 
The pattern `**/*Test.java` rather than `**/*Tests.java` so as to be 
consistent with the other applications in the project. There aren’t any
other test in this application (other than `VaultsServiceIT` which 
wasn’t picked up by the previous pattern anyway so wan’t being run) so
this doesn’t have any impact on existing tests.